requirements.txt
streamlit>=1.34.0
librosa
tensorflow-cpu
numpy==1.26.4
scikit-learn==1.8.0